+++ src/es2015/utils/helpers/drawutilssvg.js
@@ -10,7 +10,9 @@
  * @modified 2021-02-03 Fixed the currentId='background' bug on the clear() function.
  * @modified 2021-02-03 Fixed CSSProperty `stroke-width` (was line-width before, which is wrong).
  * @modified 2021-02-03 Added the static `HEAD_XML` attribute.
- * @version  1.0.0
+ * @modified 2021-02-19 Added the static helper function `transformPathData(...)` for svg path transformations (scale and translate).
+ * @modified 2021-02-22 Added the static helper function `copyPathData(...)`.
+ * @version  1.0.1
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.drawutilssvg = void 0;
@@ -815,6 +817,27 @@ class drawutilssvg {
     }
     ;
     /**
+     * Draw an SVG-like path given by the specified path data.
+     *
+     * @method path
+     * @param {SVGPathData} pathData - An array of path commands and params.
+     * @param {string=null} color - (optional) The color to draw this path with (default is null).
+     * @param {number=1} lineWidth - (optional) the line width to use (default is 1).
+     * @param {boolean=false} inplace - (optional) If set to true then path transforamtions (scale and translate) will be done in-place in the array. This can boost the performance.
+     * @instance
+     * @memberof drawutils
+     * @return {R} An instance representing the drawn path.
+     */
+    path(pathData, color, lineWidth, inplace) {
+        const node = this.makeNode('path');
+        // Transform the path: in-place (fast) or copy (slower)
+        const d = inplace ? pathData : drawutilssvg.copyPathData(pathData);
+        drawutilssvg.transformPathData(d, this.offset, this.scale);
+        node.setAttribute('d', d.join(' '));
+        return this._bindFillDraw(node, 'path', color, lineWidth);
+    }
+    ;
+    /**
      * Due to gl compatibility there is a generic 'clear' function required
      * to avoid accessing the context object itself directly.
      *
@@ -878,14 +901,40 @@ class drawutilssvg {
         return document.createElementNS("http://www.w3.org/2000/svg", "svg");
     }
     ;
+    /**
+     * Create a copy of the given path data. As path data only consists of strings and numbers,
+     * the copy will be shallow by definition.
+     *
+     * @name copyPathData
+     * @static
+     * @memberof drawutilssvg
+     */
+    static copyPathData(data) {
+        const copy = new Array(data.length);
+        for (var i = 0, n = data.length; i < n; i++) {
+            copy[i] = data[i];
+        }
+        return copy;
+    }
+    ;
+    /**
+     * Transform the given path data (translate and scale. rotating is not intended here).
+     *
+     * @name transformPathData
+     * @static
+     * @memberof drawutilssvg
+     * @param {SVGPathParams} data - The data to transform.
+     * @param {XYCoords} offset - The translation offset (neutral is x=0, y=0).
+     * @param {XYCoords} scale - The scale factors (neutral is x=1, y=1).
+     */
     static transformPathData(data, offset, scale) {
         // Scale and translate {x,y}
         const _stx = (index) => { data[index] = offset.x + scale.x * Number(data[index]); };
         const _sty = (index) => { data[index] = offset.y + scale.y * Number(data[index]); };
         // translate {x,y}
-        const _tx = (index) => { data[index] = offset.x + Number(data[index]); };
-        const _ty = (index) => { data[index] = offset.y + Number(data[index]); };
-        // scale {x,y}
+        // const _tx = (index:number) : void => { data[index] = offset.x + Number(data[index]); }
+        // const _ty = (index:number) : void => { data[index] = offset.y + Number(data[index]); }
+        // scale only {x,y}
         const _sx = (index) => { data[index] = scale.x * Number(data[index]); };
         const _sy = (index) => { data[index] = scale.y * Number(data[index]); };
         var i = 0;
@@ -913,19 +962,26 @@ class drawutilssvg {
                     i += 3;
                     break;
                 case 'H':
-                case 'h':
                     // HorizontalLineTo: H|h x
                     _stx(i + 1);
                     i += 2;
                     break;
+                case 'h':
+                    // HorizontalLineTo: H|h x
+                    _sx(i + 1);
+                    i += 2;
+                    break;
                 case 'V':
-                case 'v':
                     // VerticalLineTo: V|v y
                     _sty(i + 1);
                     i += 2;
                     break;
+                case 'v':
+                    // VerticalLineTo: V|v y
+                    _sy(i + 1);
+                    i += 2;
+                    break;
                 case 'C':
-                case 'c':
                     // CurveTo: C|c x1 y1 x2 y2 x y
                     _stx(i + 1);
                     _sty(i + 2);
@@ -935,10 +991,18 @@ class drawutilssvg {
                     _sty(i + 6);
                     i += 7;
                     break;
+                case 'c':
+                    // CurveTo: C|c x1 y1 x2 y2 x y
+                    _sx(i + 1);
+                    _sy(i + 2);
+                    _sx(i + 3);
+                    _sy(i + 4);
+                    _sx(i + 5);
+                    _sy(i + 6);
+                    i += 7;
+                    break;
                 case 'S':
-                case 's':
                 case 'Q':
-                case 'q':
                     // Shorthand-/SmoothCurveTo: S|s x2 y2 x y
                     // QuadraticCurveTo: Q|q x1 y1 x y
                     _stx(i + 1);
@@ -947,13 +1011,30 @@ class drawutilssvg {
                     _sty(i + 4);
                     i += 5;
                     break;
+                case 's':
+                case 'q':
+                    // Shorthand-/SmoothCurveTo: S|s x2 y2 x y
+                    // QuadraticCurveTo: Q|q x1 y1 x y
+                    _sx(i + 1);
+                    _sy(i + 2);
+                    _sx(i + 3);
+                    _sy(i + 4);
+                    i += 5;
+                    break;
                 case 'A':
-                case 'a':
                     // EllipticalArcTo: A|a rx ry x-axis-rotation large-arc-flag sweep-flag x y
-                    _sx(i + 1); // data[i+1] = scale.x * Number(data[i+1]);
-                    _sy(i + 2); // data[i+2] = scale.y * Number(data[i+2]);
+                    _sx(i + 1);
+                    _sy(i + 2);
                     _stx(i + 6);
                     _sty(i + 7);
+                    i += 8;
+                    break;
+                case 'a':
+                    // EllipticalArcTo: A|a rx ry x-axis-rotation large-arc-flag sweep-flag x y
+                    _sx(i + 1);
+                    _sy(i + 2);
+                    _sx(i + 6);
+                    _sy(i + 7);
                     i += 8;
                     break;
                 case 'z':
@@ -964,60 +1045,6 @@ class drawutilssvg {
                 // Safepoint: continue reading token by token until something is recognized again
                 default: i++;
             }
-            /*
-            var lastPos = [ 0, 0 ]; var pointOne, pointTwo;
-            commandList.forEach(function(command) {
-            if ((command.marker === 'z') || (command.marker === 'Z')) {
-                lastPos = [ 0, 0 ];
-                ctx.closePath();
-            } else if (command.marker === 'm') {
-                lastPos = [ lastPos[0] + command.values[0], lastPos[1] + command.values[1] ];
-                ctx.moveTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'l') {
-                lastPos = [ lastPos[0] + command.values[0], lastPos[1] + command.values[1] ];
-                ctx.lineTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'h') {
-                lastPos = [ lastPos[0] + command.values[0], lastPos[1] ];
-                ctx.lineTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'v') {
-                lastPos = [ lastPos[0], lastPos[1] + command.values[0] ];
-                ctx.lineTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'c') {
-                pointOne = [ lastPos[0] + command.values[0],
-                     lastPos[1] + command.values[1] ];
-                pointTwo = [ lastPos[0] + command.values[2],
-                     lastPos[1] + command.values[3] ];
-                lastPos  = [ lastPos[0] + command.values[4],
-                     lastPos[1] + command.values[5] ];
-                ctx.bezierCurveTo(
-                pointOne[0], pointOne[1],
-                pointTwo[0], pointTwo[1],
-                lastPos[0], lastPos[1]);
-            } else if (command.marker === 'M') {
-                lastPos = [ command.values[0], command.values[1] ];
-                ctx.moveTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'L') {
-                lastPos = [ command.values[0], lastPos[1] ];
-                ctx.lineTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'H') {
-                lastPos = [ command.values[0], lastPos[1] ];
-                ctx.lineTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'V') {
-                lastPos = [ lastPos[0], command.values[0] ];
-                ctx.lineTo(lastPos[0], lastPos[1]);
-            } else if (command.marker === 'C') {
-                pointOne = [ command.values[0],
-                     command.values[1] ];
-                pointTwo = [ command.values[2],
-                     command.values[3] ];
-                lastPos  = [ command.values[4],
-                     command.values[5] ];
-                ctx.bezierCurveTo(
-                pointOne[0], pointOne[1],
-                pointTwo[0], pointTwo[1],
-                lastPos[0], lastPos[1]);
-            }
-            */
         } // END while
     }
     ;